+++ compiler-tests/src/test/data/diagnostic/inject/assisted/AssistedFactoryExtraParamReportsOnParam.kt
@@ -1,11 +1,11 @@
 // RENDER_DIAGNOSTICS_FULL_TEXT
+
 @AssistedInject
 class ExampleClass(
   @Assisted val count: Int,
-  @Assisted val message: String,
 )
 
 @AssistedFactory
 fun interface ExampleClassFactory {
-  fun <!ASSISTED_INJECTION_ERROR!>create<!>(count: Int): ExampleClass
+  fun create(count: Int, <!ASSISTED_INJECTION_ERROR!>message<!>: String): ExampleClass
 }