+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -980,6 +980,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
       }
 
       @Test
+      @TestMetadata("AssistedFactoryExtraParamReportsOnParam.kt")
+      public void testAssistedFactoryExtraParamReportsOnParam() {
+        run("AssistedFactoryExtraParamReportsOnParam.kt");
+      }
+
+      @Test
       @TestMetadata("AssistedFactoryMustTargetAssistedInjectTypesButMissingConstructor.kt")
       public void testAssistedFactoryMustTargetAssistedInjectTypesButMissingConstructor() {
         run("AssistedFactoryMustTargetAssistedInjectTypesButMissingConstructor.kt");
@@ -1013,6 +1019,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
       @TestMetadata("ContributesAnnotationsNotAllowedOnAssistedInjectClasses.kt")
       public void testContributesAnnotationsNotAllowedOnAssistedInjectClasses() {
         run("ContributesAnnotationsNotAllowedOnAssistedInjectClasses.kt");
+      }
+
+      @Test
+      @TestMetadata("InheritedAssistedFactoryParamMismatchReportsOnFactory.kt")
+      public void testInheritedAssistedFactoryParamMismatchReportsOnFactory() {
+        run("InheritedAssistedFactoryParamMismatchReportsOnFactory.kt");
       }
 
       @Test